♻️ : FeedModal Components Refactoring
