+++ src/components/common/feedItem/indexProfile.tsx
@@ -2,8 +2,8 @@ import styled from "styled-components";
 import { useState } from "react";
 import ReactPlayer from "react-player";
 
-import FeedModalProfile from "./feedModal/indexProfile";
 import { curTabType } from "../../../types/profileType";
+import FeedModal from "./feedModal";
 
 interface feedItemType {}
 
@@ -43,7 +43,7 @@ const FeedItemProfile = ({
         )}
       </FeedItemDiv>
       {feedModal ? (
-        <FeedModalProfile
+        <FeedModal
           postType={postType}
           feedIdProfile={feedId}
           modalState={feedModal}
